+++ src/options-validation/keyword-props.js
@@ -1,0 +1,43 @@
+const validationTasks = require("../common/validation-tasks");
+const numberLimits = require("../common/number-limits");
+
+
+function validateKeywordProperties(optionsObject, propName, resultObject)
+{
+	var keywordObject = optionsObject[propName];
+	
+	handleChanceProperty(keywordObject, propName, resultObject);
+	handleKeywordProperty(keywordObject, propName, "minKeywords", resultObject);
+	handleKeywordProperty(keywordObject, propName, "maxKeywords", resultObject);
+}
+
+
+function handleChanceProperty(keywordObj, parentName, resObject)
+{
+	var givenValue = keywordObj.chance;
+	var nestString = parentName + ".chance";
+	
+	if (resObject.valid === true)
+	{
+		validationTasks.checkPercentage(givenValue, nestString, resObject);
+	}
+}
+
+
+function handleKeywordProperty(keywordObj, parentName, childName, resObject)
+{
+	var givenValue = keywordObj[childName];
+	var nestString = [parentName, childName].join(".");
+	
+	if (resObject.valid === true)
+	{
+		validationTasks.checkRange(givenValue, nestString, numberLimits.keywords, resObject);
+	}
+}
+
+
+
+module.exports =
+{
+	validateKeywords: validateKeywordProperties
+};